escrow_backend_full (1)/requirements.txt
Django==5.0.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
drf-spectacular-sidecar==2024.1.1
graphene-django==3.2.2
channels==4.1.0
web3==6.20.1
psycopg2-binary==2.9.9
